refactor(file-exchange): extract shared staging primitives (#146)
